Put the secrets on kubernetes in a secret (#1131)
Putting secrets in a secret in kubernetes makes the
intention more clear.
Also in EKS for example secrets can be encrypted at
rest thus it is best practice to use secrets for sensitive
values.

I also added a checksum of the configmap and secret to
the deployment, this makes sure that the new pods
are deployed when users change the config of their
installation