Remove popover from cart button
It was confusing that clicking on the cart
brought you to checkout, but also it had
the popover.

After some investigation we noticed that 2 out
of the 3 online stores we tested just bring you
to the checkout page.
And only one had the popover, but on click
instead. Which meant more clicks, so decided
to reduce clicks to 1 and bring you to check out
in one click.

In support of: #1917